refactor: renamed components
