+++ src/__ tests __/Footer.spec.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render, screen, fireEvent } from '@testing-library/react';
-import Footer from '../components/Footer';
+import Footer from '../components/Footer/Footer';
 
 test('link opens the correct URL', () => {
   render(<Footer />);
