+++ src/__ tests __/pagination.spec.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { fireEvent, render } from '@testing-library/react';
-import Pagination from '../components/Pagination';
+import Pagination from '../components/Pagination/Pagination';
 
 describe('Pagination component', () => {
   const onPageChangeMock = jest.fn();
