+++ src/components/CoinCard/CoinCard.tsx
