+++ src/main.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './components/App/index.tsx';
+import App from './components/App';
 import { Provider } from 'react-redux';
 import { store } from './redux';
-import './index.scss';
-import ErrorBoundary from './components/ErrorBoundary/index.tsx';
+import ErrorBoundary from './components/ErrorBoundary';
 import { ThemeProvider } from './context/ThemeProvider.tsx';
+import './index.scss';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
