+++ src/pages/Error404/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Error404';
